Remove height from test-html for compatibility
The height as caculated varies. I get 42 on Firefox and with qunit on
the command-line of my local machine, 40 on Chromium and 43 with qunit
on the command-line at travis-ci (see
https://travis-ci.org/cosmocode/dokuwiki-plugin-tablelayout/builds/164004562 ).

So this now extracts the height which is expected to be two digits and
tests it and the remaining html separately. This solution is not the
nicest, but it works for now.