Clear Floats at the end of each section
This is done to prevent a right floating table from floating into the
next section. It also ensures that the highlight of the editbutton still
works correctly.

Known issues introduced by this change:
--------------------------------------
Now the TOC is also cleared before the second headline. A way to handle
this would be to either clear floats only manually by using for example
the wrap plugin. Or to set the clear after by JS only for those sections
that contain right floating tables. However that would cause more
complexity.