Centered tables should not be floated next to others
If a previous table would be set to floating left or right, then a
following centered table might be float next to that table. That is not
supposed to happen, because a centered table is supposed to not have
text/elements to the left or right.
This commit fixes that behaviour.